requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
Pillow
numpy
opencv-python-headless
scipy
PyMuPDF
tensorflow
streamlit
streamlit-drawable-canvas
pymongo
gridfs